+++ Classes/Service/ProcessService.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace AOE\Crawler\Service;
 
 /*
- * (c) 2021 Tomas Norre Mikkelsen <tomasnorre@gmail.com>
+ * (c) 2022 Tomas Norre Mikkelsen <tomasnorre@gmail.com>
  *
  * This file is part of the TYPO3 Crawler Extension.
  *
@@ -19,15 +19,14 @@ namespace AOE\Crawler\Service;
  * The TYPO3 project - inspiring people to share!
  */
 
-use AOE\Crawler\Configuration\ExtensionConfigurationProvider;
 use AOE\Crawler\Domain\Repository\ProcessRepository;
 use AOE\Crawler\Exception\ProcessException;
+use AOE\Crawler\Helper\Sleeper\SleeperInterface;
 use AOE\Crawler\Utility\PhpBinaryUtility;
 use TYPO3\CMS\Core\Core\Environment;
 use TYPO3\CMS\Core\Information\Typo3Version;
 use TYPO3\CMS\Core\Utility\CommandUtility;
 use TYPO3\CMS\Core\Utility\ExtensionManagementUtility;
-use TYPO3\CMS\Core\Utility\GeneralUtility;
 
 /**
  * @package AOE\Crawler\Service
@@ -37,35 +36,25 @@ use TYPO3\CMS\Core\Utility\GeneralUtility;
  */
 class ProcessService
 {
-    /**
-     * @var int
-     */
-    private $timeToLive;
+    private ProcessRepository $processRepository;
 
-    /**
-     * @var \AOE\Crawler\Domain\Repository\ProcessRepository
-     */
-    private $processRepository;
+    private SleeperInterface $sleeper;
 
-    /**
-     * @var array
-     */
-    private $extensionSettings;
-
-    public function __construct()
-    {
-        $this->processRepository = GeneralUtility::makeInstance(ProcessRepository::class);
-        $this->extensionSettings = GeneralUtility::makeInstance(ExtensionConfigurationProvider::class)->getExtensionConfiguration();
-        $this->timeToLive = (int) $this->extensionSettings['processMaxRunTime'];
+    public function __construct(
+        ProcessRepository $processRepository,
+        SleeperInterface $sleeper
+    ) {
+        $this->processRepository = $processRepository;
+        $this->sleeper = $sleeper;
     }
 
     /**
      * starts new process
      * @throws ProcessException if no crawler process was started
      */
-    public function startProcess(): bool
+    public function startProcess(int $timeToLive = 300): bool
     {
-        $ttl = (time() + $this->timeToLive - 1);
+        $ttl = (time() + $timeToLive - 1);
         $current = $this->processRepository->countNotTimeouted($ttl);
 
         // Check whether OS is Windows
@@ -85,7 +74,7 @@ class ProcessService
             if ($this->processRepository->countNotTimeouted($ttl) > $current) {
                 return true;
             }
-            sleep(1);
+            $this->sleeper->sleep(1);
         }
         throw new ProcessException('Something went wrong: process did not appear within 10 seconds.');
     }
@@ -113,11 +102,6 @@ class ProcessService
         }
 
         return ltrim($scriptPath);
-    }
-
-    public function setProcessRepository(ProcessRepository $processRepository): void
-    {
-        $this->processRepository = $processRepository;
     }
 
     private function getComposerBinPath(): ?string